+++ Server/Utils/anteriorNodemailer.js
@@ -1,0 +1,13 @@
+const nodemailer = require('nodemailer')
+
+const transporter = nodemailer.createTransport({
+    host: 'smtp.gmail.com',
+    port: 465,
+    secure: true,
+    auth: {
+        user: process.env.EMAIL,
+        pass: process.env.EMAIL_PASSWORD,
+    },
+})
+
+module.exports = transporter
